Yawn, I looked outside it's still winter.
I have been thinking about trying something new with my website. Try bringing it to life. My site was originally a sort of Almanac. Mostly Astronomical and Astrological calculations.  My database I use it in a strange way. The data the output is expressed using a, what is normally an input. I put the calculations into tables in Excel then into Access.  The positions in the sky of the sun, moon, planets, asteroids, and most if not all named stars. I enter the positions of the moon. The phases of the moon lunar days,  the lunar Mansions Void of course of the moon. The planetary hours and days.  Boring stuff has been instead of the daily forecast of the moon position. Coming out of the database in text, and having to read it. Thought of using text to speech technology to render it into speech. Try to bring my own interest in it all back. I used to make posters, I would basically give them away in PDF.  Well, I am not much of a promoter or salesman.  I never had any real training in database design. To someone who actually knows what they are doing, I could imagine just how ridiculous and backward the way I do things would appear.  Hello